Moving decentraland/screen/helpers.js -> decentraland/helpers.js
